# backend/novu-bridge/src/main/resources/contract/openapi.yaml
openapi: 3.0.3
info:
  title: novu-bridge — notification delivery adapter
  version: "1"
  description: |
    Every HTTP endpoint novu-bridge exposes, as implemented. Written from the controllers in
    `backend/novu-bridge/src/main/java/org/egov/novubridge/web/controllers/`; `OpenApiContractTest`
    fails the build if a path here has no matching controller mapping, or a controller mapping
    is missing here.

    The bridge's PRIMARY interface is not HTTP at all — it is the pair of Kafka messages
    described by `envelope-v1.schema.json` (a finished message) and `thin-event-v1.schema.json`
    (a domain event the box resolves), told apart by the `kind` discriminator. These endpoints
    are the management and observability surface the Configurator drives, plus machine
    callbacks.

    ## Tiers

    | Tier | Endpoints | Who may call |
    |---|---|---|
    | Operator (read + exercise) | `/config/source`, `/logs`, `/integrations`, `/preferences`, `/providers/catalog`, `/providers/templates`, `/providers/twilio-templates`, `/providers/verify`, `/providers/test-send` | DIGIT bearer token, `type == EMPLOYEE`, a role from `novu.bridge.proxy.allowed.roles` (or from the admin list, which is a superset by intent) |
    | Admin (credentials) | `POST /providers`, `POST /providers/_update`, `POST /providers/_delete`, `POST /dispatch/_resolve` | the above **and** a role from `novu.bridge.proxy.admin.roles`; otherwise `403 NB_ADMIN_ROLE_REQUIRED` |
...

...
    unauthenticated by design.

...

...
servers:
...
security:
  - digitBearer: []
paths:
  /novu-bridge/novu-adapter/v1/contract/envelope:
    get:
      tags: [contract]
      summary: JSON Schema of the inbound Kafka envelope (version 1)
      description: |
        The byte-identical copy packaged in this build's jar, so what you fetch is what this
        build enforces. Unauthenticated: a description of an interface, not of a deployment.
      security: []
      responses:
        "200":
          description: JSON Schema 2020-12 document
          content:
            application/json:
              schema:
                type: object
        "404":
          description: The jar was built without `contract/envelope-v1.schema.json`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /novu-bridge/novu-adapter/v1/contract/thin-event:
    get:
      tags: [contract]
      summary: JSON Schema of the inbound thin domain event (version 1)
      description: |
        The second inbound kind: a producer says only what happened to which entity, and the
        box decides who to tell, on which channel, in which language and what the words are.
        Told apart from the envelope by the `kind` discriminator (`THIN`); see `README.md`.

        The byte-identical copy packaged in this build's jar. Unauthenticated, for the same
        reason as the other two: a description of an interface, not of a deployment.
      security: []
      responses:
        "200":
          description: JSON Schema 2020-12 document
          content:
            application/json:
              schema:
                type: object
        "404":
          description: The jar was built without `contract/thin-event-v1.schema.json`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /novu-bridge/novu-adapter/v1/contract/openapi:
    get:
      tags: [contract]
      summary: This document
      security: []
      responses:
        "200":
          description: OpenAPI 3.0 document
          content:
            application/yaml:
              schema:
                type: string
        "404":
          description: The jar was built without `contract/openapi.yaml`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /novu-bridge/novu-adapter/v1/dispatch/_resolve:
    post:
      tags: [dispatch]
      summary: Resolve a thin domain event WITHOUT sending or writing a row
      description: |
        Takes a thin domain event and returns the v1 envelopes the box WOULD mint for it: one per
        recipient x channel, in emission order, fully rendered, with the contact block filled in.
        Nothing is dispatched and **no ledger row is written**, which is what makes it safe to
        point at a production tenant.

        It runs the real resolver against the tenant's real config — the same routing rows,
        templates, role pools and locale preferences — so it answers "why would this event send
        nothing?" before the event happens rather than forensically afterwards. `terminalCode`
        carries the channel-less decision when there would be none (`NB_NO_ROUTING`,
        `NB_NO_RECIPIENTS`, `NB_UNKNOWN_AUDIENCE_SCHEME`, `NB_RECIPIENT_LIMIT_EXCEEDED`) and
        `diagnostics` lists every decision taken on the way, including per-recipient skips.

        **Admin-only.** It expands role pools and returns recipient PII for every holder of a
        role, which is broader than the Logs screen's read tier should hand out.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ThinEventResolveRequest" }
      responses:
        "200":
          description: The envelopes the event would produce
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ThinEventResolveResponse" }
        "400":
          description: "`NB_EVENT_NOT_IN_CATALOGUE` — the event name has no active catalogue row for this tenant."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /novu-bridge/novu-adapter/v1/dispatch/_validate:
    post:
      tags: [dispatch]
      summary: Run one envelope through the pipeline WITHOUT sending
      description: |
        Diagnostic. Validates the envelope, derives the context and persists a `RECEIVED` row.
        Nothing is handed to a provider.

        **Not routed publicly by Kong.** Reachable in-cluster; inside the service it is gated
        by `ProxyAuthFilter` like the other `/dispatch` paths.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DispatchDryRunRequest" }
      responses:
        "200":
          description: The pipeline's verdict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchDryRunResponse" }
        "400":
          description: "Envelope rejection — `NB_INVALID_EVENT`, `NB_UNSUPPORTED_SCHEMA_VERSION`, `NB_UNSUPPORTED_EVENT_TYPE`. A `REJECTED` row is written before the error is returned."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /novu-bridge/novu-adapter/v1/dispatch/_dry-run:
    post:
      tags: [dispatch]
      summary: "Same as _validate, but `send: true` actually delivers"
      description: |
        Diagnostic. With `send` absent or false this behaves exactly like `_validate`. With
        `send: true` the envelope goes down the full live path, gates and all, and a real
        message is delivered. **Not routed publicly by Kong.**
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DispatchDryRunRequest" }
      responses:
        "200":
          description: The pipeline's verdict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DispatchDryRunResponse" }
        "400":
          description: "Envelope rejection. A `REJECTED` row is written before the error is returned."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    digitBearer:
      type: http
      scheme: bearer
      description: |
        A DIGIT access token, introspected by the bridge against egov-user
        `POST /user/_details`. The caller must be `type == EMPLOYEE` and hold a role from
        `novu.bridge.proxy.allowed.roles`; the three credential-bearing provider calls need a
        role from `novu.bridge.proxy.admin.roles` as well.
    receiptSecret:
      type: apiKey
      in: header
      name: X-Receipt-Secret
      description: "`novu.bridge.receipts.secret`. May also be sent as the `secret` query parameter."
    smsCountryUser:
      type: apiKey
      in: header
      name: X-SMSCountry-User
    smsCountryPassword:
      type: apiKey
      in: header
      name: X-SMSCountry-Password
  responses:
    Unauthorized:
      description: No bearer token, or the token does not introspect
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: string } }
    Forbidden:
      description: The caller is not an EMPLOYEE, or holds no allowlisted role
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: string } }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string, example: NB_NO_PROVIDER }
        message: { type: string }
        description: { type: string }
    ErrorEnvelope:
      type: object
      description: The shared DIGIT error shape rendered by the tracer handler.
      properties:
        ResponseInfo: { type: object, nullable: true }
        Errors:
          type: array
          items: { $ref: "#/components/schemas/Error" }
...
    NotificationEvent:
      type: object
      description: |
        The inbound Kafka envelope. Authoritatively defined by `envelope-v1.schema.json` —
        this entry exists only so the `/dispatch` diagnostics can reference it.
      externalDocs:
        description: envelope-v1.schema.json
        url: ./envelope-v1.schema.json
      additionalProperties: true
...
    ThinEventResolveRequest:
      type: object
      required: [RequestInfo, event]
      properties:
        RequestInfo:
          type: object
          description: The standard DIGIT RequestInfo.
        event:
          type: object
          description: "A thin domain event; see `contract/thin-event-v1.schema.json`."
    ThinEventResolveResponse:
      type: object
      properties:
        responseInfo: { type: object }
        envelopes:
          type: array
          description: One per recipient x channel, in emission order. Empty when nothing would be sent.
          items: { $ref: "#/components/schemas/NotificationEvent" }
        terminalCode:
          type: string
          nullable: true
          description: "The channel-less decision code when nothing would be sent — `NB_NO_ROUTING`, `NB_NO_RECIPIENTS`, `NB_UNKNOWN_AUDIENCE_SCHEME`, `NB_RECIPIENT_LIMIT_EXCEEDED`."
        diagnostics:
          type: array
          items: { type: string }
          description: Every decision the resolver took, in order, including per-recipient skips.
    DispatchDryRunRequest:
      type: object
      required: [RequestInfo, event]
      properties:
        RequestInfo:
          type: object
          description: The standard DIGIT RequestInfo.
        send:
          type: boolean
          default: false
          description: "`_dry-run` only. True delivers for real."
        event: { $ref: "#/components/schemas/NotificationEvent" }
    DispatchDryRunResponse:
      type: object
      properties:
        responseInfo: { type: object }
        result:
          type: object
          properties:
            valid: { type: boolean }
            preferenceAllowed: { type: boolean }
            derivedContext:
              type: object
              description: What the bridge read out of the envelope — channel, audience, locale, recipient, subscriberId, body, subject, transactionId.
            novuTriggered: { type: boolean }
            novuStatusCode: { type: integer, nullable: true }
            novuResponse: { type: object, nullable: true, additionalProperties: true }
            diagnostics:
              type: array
              items: { type: string }
              description: One line saying what happened, e.g. "Preference denied".